Quick fix for Issue #2 (Doesn't expand all goals)
This adds the goal to the __aux fact that tracks whether
the goal has been expanded or not.

This covers most cases of expanding multiple goals in
a term, but will fail to expand to identical goals
(same predicate and args).

This has been tested by hand but there is no
automated test for it yet.